+++ ClubManagerWeb/src/main/java/com/infoshareacademy/niewiem/services/TableDeleteService.java
@@ -1,6 +1,7 @@
 package com.infoshareacademy.niewiem.services;
 
 import com.infoshareacademy.niewiem.dao.TableDao;
+import com.infoshareacademy.niewiem.services.validators.TableValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,10 +15,18 @@ public class TableDeleteService {
     @Inject
     private TableDao tableDao;
 
+    @Inject
+    private TableValidator tableValidator;
+
     public void delete(Integer id) {
+
+        if(tableValidator.isIdNotNull(id)){
+            tableDao.delete(id);
+        }
         // todo: validate me like you validate your French girls!
         // Should deleting table cancel all future reservations?
         // Maybe it should send an email to admin with reservations and customer contact to inform about cancelation.
-        tableDao.delete(id);
+        LOG.warn("Table didn't delete because id is null");
+        throw new IllegalArgumentException("Table with id '" + id + "' doesn't exist");
     }
 }
